# swagger.yaml
swagger: '2.0'
info:
  description: Набор сервисов для работы с системой учета /ООО Ромашка/
  version: "1.0.0"
  title: RestAPI
consumes:
  - application/json
  - application/xml
  - application/text
produces:
  - application/json  
  - application/xml
  - application/text  
basePath: "/api"   

tags:
  - name: Reports
    description: Набор методов для работы с отчетностью
  - name: Dictionary
    description: Справочники  

paths:
  /dictionary/formats:
    get:
      tags:
        - Dictionary
      summary: Получить список форматов
      operationId: main.formats
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере    
        default:
          description: Неизвестная ошибка      
           
  /dictionary/entities:
    get:
      tags:
        - Dictionary
      summary: Получить список моделей данных
      operationId: main.entities
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере    
        default:
          description: Неизвестная ошибка     

  /dictionary/conditions:
    get:
      tags:
        - Dictionary
      summary: Получить список вариантов сравнения для создания фильтров
      operationId: main.conditions
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере    
        default:
          description: Неизвестная ошибка      

  /reports/{entity}/{format}:
    get:
      tags:
        - Reports
      summary: Получить данные
      operationId: main.get_report
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере    
        default:
          description: Неизвестная ошибка   
    parameters:
    - name: entity
      in: path
      description: Уникальное наименование модели данных (/dictionary/entities)
      required: true
      type: string
    - name: format
      in: path
      description: Уникальный код формата (/dictionary/formats)
      required: true
      type: integer

  /reports/{entity}:
    post:
      tags:
          - Reports
      summary: Получить данные с применением фильтрации
      operationId: main.get_filtered_report
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере    
        default:
            description: Неизвестная ошибка   
    parameters:
    - name: entity
      in: path
      description: Уникальное наименование модели данных (/dictionary/entities)
      required: true
      type: string
    - name: data  
      in: body
      description: Структура для выполнения фильтрации данных
      schema:
        properties:
          items:
            $ref: '#/definitions/filter'

  /app/date_block:
    get:
      summary: Получить текущую дату блокировки
      operationId: main.get_data_block
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    post:
      summary: Изменить дату блокировки
      operationId: main.set_data_block
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          name: data_block
          description: Новая дата блокировки
          required: true
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /app/nomenclature/get/{item_id}:
    get:
      operationId: main.nomenclature_get
      parameters:
        - name: item_id
          in: path
          required: true
          type: string
      produces:
        - application/text
      responses:
        200:
          description: OK
          schema:
            type: string
        404:
          description: "Элемент не найден"
        500:
          description: "Ошибка на сервере"

  /app/nomenclature/add:
    post:
      operationId: main.nomenclature_add
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            additionalProperties: true
      produces:
        - application/text
      responses:
        200:
          description: "Элемент успешно создан"
          schema:
            type: "object"
        400:
          description: "Некорректные данные"
        500:
          description: "Ошибка на сервере"

  /app/nomenclature/update:
    post:
      operationId: main.nomenclature_update
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "Данные для обновления элемента"
          required: true
          schema:
            type: string
            additionalProperties: true
      produces:
        - application/text
      responses:
        200:
          description: "Элемент успешно обновлен"
          schema:
            type: string
        400:
          description: "Некорректные данные"
        500:
          description: "Ошибка на сервере"

  /app/nomenclature/delete/{item_id}:
    delete:
      operationId: main.nomenclature_delete
      parameters:
        - name: item_id
          in: path
          required: true
          type: string
          description: "ID элемента номенклатуры"
      produces:
        - application/text
      responses:
        200:
          description: "Элемент успешно удален"
          schema:
            type: string
        404:
          description: "Элемент не найден"
        500:
          description: "Ошибка на сервере"
        
  /app/save_data:
    post:
      operationId: main.save_data
      consumes:
        - application/json
      produces:
        - application/text
      responses:
        200:
          description: "Элемент успешно обновлен"
          schema:
            type: "object"
        400:
          description: "Некорректные данные"
        500:
          description: "Ошибка на сервере"

  /app/upload_data:
    post:
      operationId: main.upload_data
      consumes:
        - application/json
      produces:
        - application/text
      responses:
        200:
          description: "Элемент успешно обновлен"
          schema:
            type: "object"
        400:
          description: "Некорректные данные"
        500:
          description: "Ошибка на сервере"

  /app/get_osb:
    get:
      summary: Получить Оборотно-сальдовая ведомость
      operationId: main.get_osb
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

           
definitions:
    filter:
        type: array
        example: [
                    {
                    "field": "name",
                    "condition": 2,
                    "value": "м"
                    },
                    {
                    "field": "name",
                    "condition": 2,
                    "value": "мука"
                    }
                 ]         

                  


 